requirements.txt
fastapi
uvicorn
telethon
python-dotenv
redis
